+++ include/cask/observable/FlatMapObserver.hpp
@@ -27,10 +27,10 @@ public:
     void onError(E value);
     void onComplete();
 private:
-
     std::function<ObservableRef<TO,E>(TI)> predicate;
     std::shared_ptr<Observer<TO,E>> downstream;
     std::shared_ptr<Scheduler> sched;
+    std::atomic_flag completed;
 };
 
 
@@ -42,38 +42,47 @@ FlatMapObserver<TI,TO,E>::FlatMapObserver(
     : predicate(predicate)
     , downstream(downstream)
     , sched(sched)
+    , completed(false)
 {}
 
 template <class TI, class TO, class E>
 DeferredRef<Ack,E> FlatMapObserver<TI,TO,E>::onNext(TI value) {
     return predicate(value)
-    ->template mapTask<Ack>([downstream = downstream](auto resultValue) {
-        return Task<Ack,E>::deferAction([downstream, resultValue](auto) {
-            return downstream->onNext(resultValue);
-        });
-    })
-    ->takeWhileInclusive([](auto ack){
-        return ack == Continue;
-    })
-    ->last()
-    .template map<Ack>([downstream = downstream](auto lastAckOpt) {
-        if(lastAckOpt.has_value()) {
-            return *lastAckOpt;
-        } else {
-            return Continue;
-        }
-    })
-    .run(sched);
+        ->template mapTask<Ack>([downstream = downstream](auto resultValue) {
+            return Task<Ack,E>::deferAction([downstream, resultValue](auto) {
+                return downstream->onNext(resultValue);
+            });
+        })
+        ->takeWhileInclusive([](auto ack){
+            return ack == Continue;
+        })
+        ->last()
+        .template map<Ack>([downstream = downstream](auto lastAckOpt) {
+            if(lastAckOpt.has_value()) {
+                return *lastAckOpt;
+            } else {
+                return Continue;
+            }
+        })
+        .recover([this](auto error) {
+            onError(error);
+            return Stop;
+        })
+        .run(sched);
 }
 
 template <class TI, class TO, class E>
 void FlatMapObserver<TI,TO,E>::onError(E error) {
-    return downstream->onError(error);
+    if(!completed.test_and_set()) {
+        downstream->onError(error);
+    }
 }
 
 template <class TI, class TO, class E>
 void FlatMapObserver<TI,TO,E>::onComplete() {
-    return downstream->onComplete();
+    if(!completed.test_and_set()) {
+        return downstream->onComplete();
+    }
 }
 
 }